# Avoid serializing a node twice