@tizoc 2021-04-23
@sebastiencs cool, that is a good speedup. Just one thing I would change `serialized` to something else that is more descriptive about what we really care about here (`persisted`, `saved`, `committed` or something like that). Also CI is complaining because of rustfmt, otherwise, it looks good👍